converter/requirements.txt
appdirs==1.4.4
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2.16
djangorestframework==3.13.0
ffmpeg-downloader==0.2.0
flake8==5.0.4
gunicorn==20.0.4
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
packaging==23.1
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pydub==0.25.1
pyflakes==2.5.0
python-dotenv==0.21.1
pytz==2023.3
requests==2.27.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.1
urllib3==1.26.16
zipp==3.15.0
